add account feature

+++ src/Core/Entities/Identity/RolePermission.cs
@@ -11,7 +11,7 @@ namespace Core.Entities.Identity
     public class RolePermission
     {
         public string RoleId { get; set; } = string.Empty;
-        public decimal PermissionId { get; set; }
+        public long PermissionId { get; set; }
 
         public AppRole Role { get; set; } = default!;
         public Permission Permission { get; set; }  = default!;
